Transpile @kbn/analytics with Babel (#43407)
Since this is being imported by the UI as a package, we need to transpile for the
browsers we need to support.

Fixes #42782
